Reorganized the weapons within the code so that it is easier to read. The weapons have been split up base on whether they are meant to be used in hub activities, special loot from enemies, or weapons received through the story line. For the hub section, there are subsections that separate the weapons by class, and by the chapter they should be added in.
Weapons before this update that were not specified as to what chapter they belong to or if they are meant to be dropped by enemies or received during a quest, will be located at the top of the function to ensure they are still usable. 

Added in weapons for melee class.