Added new functionality to chapter 1
Made chapter 1 check to ensure that the player has bought a weapon, leveled up a class to lvl 3, and that the class they leveled up matches a weapon in their inventory. This forces the player to interact with the search around and shop functions as well as ensuring they are powerful enough to fight chapter 2 enemies. It is more a failsafe for new players.